Clean up tooltip generators
